Searching functionality works
