+++ src/app/app.component.html
@@ -1,62 +1,3 @@
 <!--The content below is only a placeholder and can be replaced.-->
-<nav role="navigation" class="navbar navbar-default">
-  <div class="header container">
-  <ul class="nav nav-pills navbar-right">
-    <li class="active"><a data-toggle="pill" href="#home"> Home</a></li>
-    <li><a data-toggle="pill" href="#movies">Movies</a></li>
-    <li><a data-toggle="pill" href="#shows">Shows</a></li>
-    <li><a data-toggle="pill" href="#categories">Categories</a></li>
-    <li><div class="searchbox">
-      <!-- <label for="Search"></label> -->
-      <input name="search" placeholder="Search List" type="text" [(ngModel)]="query"> <!-- searchText -->
-        <button (click)="searchPosts()">Submit</button>
-    </div>
-  </li>
-  <li> <input type="text"
-              placeholder="Search List"
-              class="filter-input"
-              [(ngModel)]="searchText"></li>
-  </ul>
-  </div>
-</nav> <!-- End of navbar -->
-
-  <ul>
-    <h2>List of TV Shows</h2>
-    <li *ngFor="let show of shows">
-    <!-- | searchBy: ['data.title']: [searchText]"> -->
-
-      <div class="row container">
-        <div class="col-md-4">
-          <a href="{{show.url}}" target="_blank">
-            {{show.title}}
-            <img src={{show.thumbnail}} alt="">
-          </a>
-          </div>
-          <div class="col-md-4">
-            <p class="hidden"></p>
-            <p class="hidden"></p>
-            <span>Description: {{show.description}}</span>
-            <span>Number of Episodes: {{show.length}}</span>
-            <span>Release Date: {{show.releasedate | date}}</span>
-          </div>
-        </div>
-    </li>
-    <h2>List of Movies</h2>
-    <li *ngFor="let show of movies">
-      <div class="row container">
-        <div class="col-md-4">
-          <a href="{{show.url}}" target="_blank">
-            {{show.title}}
-            <img src={{show.thumbnail}} alt="">
-          </a>
-          </div>
-          <div class="col-md-4">
-            <p class="hidden"></p>
-            <p class="hidden"></p>
-            <span>Description: {{show.description}}</span>
-            <span>Number of Episodes: {{show.length}}</span>
-            <span>Release Date: {{show.releasedate | date}}</span>
-          </div>
-        </div>
-    </li>
-  </ul>
+<navbar></navbar>
+<router-outlet></router-outlet>
