+++ src/app/homepage/homepage.component.css
@@ -1,0 +1,4 @@
+.format{
+  display: block;
+  text-align: right;
+}
