+++ src/app/homepage/homepage.component.html
@@ -1,0 +1,55 @@
+
+<!-- <div class="searchbox"> -->
+  <!-- <label for="Search"></label> -->
+  <!-- <input name="search" placeholder="Search List" type="text" [(ngModel)]="query">  -->
+  <!-- searchText -->
+    <!-- <button (click)="searchPosts()">Submit</button> -->
+<!-- </div> -->
+<div class="format">
+  <input type="text"
+          placeholder="Search List"
+          class="filter-input"
+          [(ngModel)]="searchText">
+</div>
+<div class="jumbotron container-fluid">
+  <h2>List of All TV Shows and Movies</h2>
+</div>
+  <ul>
+    <li *ngFor="let show of shows | searchBy: ['title']: [searchText]">
+    <!-- | searchBy: ['data.title']: [searchText]"> -->
+      <div class="row container">
+        <div class="col-md-4">
+          <a href="{{show.url}}" target="_blank">
+            {{show.title}}</a>
+            <img src="{{show.thumbnail}}" alt="">
+          </div>
+          <div class="col-md-8">
+            <p class="hidden"></p>
+            <p class="hidden"></p>
+            <span><strong> Description:</strong> {{show.description}}</span>
+            <br>
+            <span><strong>Number of Episodes:</strong> {{show.length}}</span>
+            <br>
+            <span><strong>Release Date:</strong> {{show.releasedate}}</span>
+          </div>
+        </div>
+    </li>
+    <li *ngFor="let show of movies | searchBy: ['title']: [searchText]">
+      <div class="row container">
+        <div class="col-md-4">
+          <a href="{{show.url}}" target="_blank">
+            {{show.title}}</a>
+            <img src={{show.thumbnail}} alt="">
+          </div>
+          <div class="col-md-8">
+            <p class="hidden"></p>
+            <p class="hidden"></p>
+            <span><strong> Description:</strong> {{show.description}}</span>
+            <br>
+            <span><strong>Number of Episodes:</strong> {{show.length}}</span>
+            <br>
+            <span><strong>Release Date:</strong> {{show.releasedate}}</span>
+          </div>
+        </div>
+    </li>
+  </ul>
