+++ src/app/navbar/navbar.component.html
@@ -1,0 +1,15 @@
+<!-- <ul class="nav nav-tabs">
+  <li role="presentation" class="active"><a href="home">Home</a></li>
+  <li role="presentation"><a href="#">Profile</a></li>
+  <li role="presentation"><a href="#">Messages</a></li>
+</ul> -->
+<nav role="navigation" class="navbar navbar-default">
+  <div class="header container-fluid">
+    <ul class="nav nav-tabs">
+      <li role="presentation"><a routerLink="/home" routerLinkActive="active"> Home</a></li>
+      <li><a data-toggle="tabs" href="movies">Movies</a></li>
+      <li><a data-toggle="tabs" href="shows">Shows</a></li>
+      <li><a data-toggle="tabs" href="categories">Categories</a></li>
+    </ul>
+  </div>
+</nav> <!-- End of navbar -->
